+++ resources/views/grid/inline-edit/select.blade.php
@@ -1,58 +1,37 @@
 @extends('admin::grid.inline-edit.comm')
 
-<span class="ie-wrap">
-    <span class="ie-display">{{ $value }}</span>
-    &nbsp;
-    <a
-        href="javascript:void(0);"
-        class="{{ $trigger }} text-muted"
-        style="visibility: hidden;"
-        data-target="{{ $target }}"
-        data-toggle="popover"
-        data-value="{{ $value }}"
-        data-original="{{ $value }}"
-    >
-        <i class="fa fa-edit"></i>
-    </a>
+@section('field')
+    <select name='select-{{ $name }}' class="form-control ie-input">
+    @foreach($options as $option => $label)
+        <option name='select-{{ $name }}' value="{{ $option }}" data-label="{{ $label }}">&nbsp;{{$label}}&nbsp;&nbsp;</option>
+    @endforeach
+    </select>
+@endsection
 
-    @component('admin::grid.inline-edit.partials.template', compact('name', 'target', 'key'))
-        <select name='radio-{{ $name }}' class="form-control ie-input" style="z-index: 1070;">
-        @foreach($options as $option => $label)
-            <option name='radio-{{ $name }}' value="{{ $option }}" data-label="{{ $label }}">&nbsp;{{$label}}&nbsp;&nbsp;</option>
-        @endforeach
-        </select>
-    @endcomponent
-</span>
+@section('assert')
+    <script>
+        @component('admin::grid.inline-edit.partials.popover', compact('trigger'))
+            @slot('popover')
+            $popover.find('select>option').each(function (index, option) {
+                if ($(option).attr('value') == $trigger.data('value')) {
+                    $(option).prop('selected', true);
+                }
+            });
+            @endslot
+        @endcomponent
+    </script>
 
-<style>
-    tbody tr:hover .{{ $trigger }}  {
-        visibility: visible !important;
-    }
-</style>
+    <script>
+    @component('admin::grid.inline-edit.partials.submit', compact('resource', 'name'))
 
-<script>
-    @component('admin::grid.inline-edit.partials.popover', compact('trigger'))
-        @slot('popover')
-        $popover.find('select>option').each(function (index, option) {
-            if ($(option).attr('value') == $trigger.data('value')) {
-                $(option).prop('selected', true);
-            }
-        });
-        $popover.find('.ie-input').select2();
+        @slot('val')
+            var val = $popover.find('.ie-input').val();
+            var label = $popover.find('.ie-input>option:selected').data('label');
         @endslot
+
+        $popover.data('display').html(label);
+
     @endcomponent
-</script>
-
-<script>
-@component('admin::grid.inline-edit.partials.submit', compact('resource', 'name'))
-
-    @slot('val')
-        var val = $popover.find('.ie-input').val();
-        var label = $popover.find('.ie-input:selected').data('label');
-    @endslot
-
-    $popover.data('display').html(label);
-
-@endcomponent
-</script>
+    </script>
+@endsection
